# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  zip?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateClientRequest {
  name: string;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  city?: string | null;
  state?: string | null;
  zip?: string | null;
  notes?: string | null;
}

export interface Estimate {
  id: number;
  estimateNumber: string;
  clientId?: number | null;
  clientName?: string | null;
  projectName: string;
  /** draft | sent | approved | rejected | invoiced */
  status: string;
  subtotal: number;
  materialsCost: number;
  laborCost: number;
  markupPercent: number;
  markupAmount: number;
  taxPercent: number;
  taxAmount: number;
  total: number;
  profitMargin: number;
  notes?: string | null;
  validUntil?: string | null;
  createdAt: string;
}

export interface LineItem {
  id: number;
  estimateId: number;
  /** labor | material | subcontractor | other */
  type: string;
  category?: string | null;
  description: string;
  quantity: number;
  unit?: string | null;
  unitCost: number;
  total: number;
  materialId?: number | null;
  supplierId?: number | null;
  supplierName?: string | null;
  notes?: string | null;
  sortOrder: number;
}

export type EstimateDetail = Estimate & {
  lineItems: LineItem[];
  client?: Client | null;
};

export interface CreateLineItemRequest {
  type?: string;
  category?: string | null;
  description: string;
  quantity: number;
  unit?: string | null;
  unitCost: number;
  materialId?: number | null;
  supplierId?: number | null;
  supplierName?: string | null;
  notes?: string | null;
  sortOrder?: number;
}

export interface CreateEstimateRequest {
  clientId?: number | null;
  projectName: string;
  status?: string;
  markupPercent?: number;
  taxPercent?: number;
  notes?: string | null;
  validUntil?: string | null;
  lineItems?: CreateLineItemRequest[];
}

export interface UpdateEstimateRequest {
  clientId?: number | null;
  projectName?: string;
  status?: string;
  markupPercent?: number;
  taxPercent?: number;
  notes?: string | null;
  validUntil?: string | null;
  lineItems?: CreateLineItemRequest[];
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  estimateId?: number | null;
  clientId?: number | null;
  clientName?: string | null;
  projectName: string;
  /** draft | sent | paid | overdue | cancelled */
  status: string;
  subtotal: number;
  markupAmount: number;
  taxAmount: number;
  total: number;
  amountPaid: number;
  amountDue: number;
  dueDate?: string | null;
  paidAt?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type InvoiceDetail = Invoice & {
  lineItems: LineItem[];
  client?: Client | null;
};

export interface CreateInvoiceRequest {
  estimateId?: number | null;
  clientId?: number | null;
  projectName: string;
  status?: string;
  dueDate?: string | null;
  notes?: string | null;
}

export interface UpdateInvoiceRequest {
  status?: string;
  dueDate?: string | null;
  notes?: string | null;
  amountPaid?: number;
  paidAt?: string | null;
}

export interface Material {
  id: number;
  name: string;
  description?: string | null;
  category?: string | null;
  unit?: string | null;
  currentPrice: number;
  supplierId?: number | null;
  supplierName?: string | null;
  sku?: string | null;
  url?: string | null;
  lastUpdated: string;
}

export interface CreateMaterialRequest {
  name: string;
  description?: string | null;
  category?: string | null;
  unit?: string | null;
  currentPrice: number;
  supplierId?: number | null;
  sku?: string | null;
  url?: string | null;
}

export interface PriceHistory {
  id: number;
  materialId: number;
  price: number;
  recordedAt: string;
  notes?: string | null;
}

export interface Supplier {
  id: number;
  name: string;
  website?: string | null;
  phone?: string | null;
  notes?: string | null;
}

export interface CreateSupplierRequest {
  name: string;
  website?: string | null;
  phone?: string | null;
  notes?: string | null;
}

export interface DashboardStats {
  totalEstimates: number;
  pendingEstimates: number;
  approvedEstimates: number;
  totalInvoices: number;
  paidInvoices: number;
  overdueInvoices: number;
  totalRevenue: number;
  pendingRevenue: number;
  totalProfit: number;
  avgProfitMargin: number;
}

export type ListEstimatesParams = {
  status?: string;
  clientId?: number;
};

export type ListInvoicesParams = {
  status?: string;
};

export type ListMaterialsParams = {
  category?: string;
  supplierId?: number;
};