# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contractor Estimate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /estimates:
    get:
      operationId: listEstimates
      tags: [estimates]
      summary: List all estimates
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Estimate"
    post:
      operationId: createEstimate
      tags: [estimates]
      summary: Create an estimate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstimateRequest"
      responses:
        "201":
          description: Created estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Estimate"
  /estimates/{id}:
    get:
      operationId: getEstimate
      tags: [estimates]
      summary: Get an estimate with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Estimate with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateDetail"
    put:
      operationId: updateEstimate
      tags: [estimates]
      summary: Update an estimate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEstimateRequest"
      responses:
        "200":
          description: Updated estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateDetail"
    delete:
      operationId: deleteEstimate
      tags: [estimates]
      summary: Delete an estimate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /estimates/{id}/convert-to-invoice:
    post:
      operationId: convertEstimateToInvoice
      tags: [estimates]
      summary: Convert estimate to invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /estimates/{id}/line-items:
    post:
      operationId: addLineItem
      tags: [estimates]
      summary: Add a line item to an estimate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLineItemRequest"
      responses:
        "201":
          description: Created line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineItem"
  /estimates/{id}/line-items/{lineItemId}:
    put:
      operationId: updateLineItem
      tags: [estimates]
      summary: Update a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: lineItemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLineItemRequest"
      responses:
        "200":
          description: Updated line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineItem"
    delete:
      operationId: deleteLineItem
      tags: [estimates]
      summary: Delete a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: lineItemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateClientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
    Estimate:
      type: object
      properties:
        id:
          type: integer
        estimateNumber:
          type: string
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        projectName:
          type: string
        status:
          type: string
          description: "draft | sent | approved | rejected | invoiced"
        subtotal:
          type: number
        materialsCost:
          type: number
        laborCost:
          type: number
        markupPercent:
          type: number
        markupAmount:
          type: number
        taxPercent:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        profitMargin:
          type: number
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - estimateNumber
        - projectName
        - status
        - subtotal
        - materialsCost
        - laborCost
        - markupPercent
        - markupAmount
        - taxPercent
        - taxAmount
        - total
        - profitMargin
        - createdAt
    EstimateDetail:
      allOf:
        - $ref: "#/components/schemas/Estimate"
        - type: object
          properties:
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/LineItem"
            client:
              $ref: "#/components/schemas/Client"
              nullable: true
          required:
            - lineItems
    CreateEstimateRequest:
      type: object
      properties:
        clientId:
          type: integer
          nullable: true
        projectName:
          type: string
        status:
          type: string
          default: draft
        markupPercent:
          type: number
          default: 0
        taxPercent:
          type: number
          default: 0
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateLineItemRequest"
      required:
        - projectName
    UpdateEstimateRequest:
      type: object
      properties:
        clientId:
          type: integer
          nullable: true
        projectName:
          type: string
        status:
          type: string
        markupPercent:
          type: number
        taxPercent:
          type: number
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/CreateLineItemRequest"
    LineItem:
      type: object
      properties:
        id:
          type: integer
        estimateId:
          type: integer
        type:
          type: string
          description: "labor | material | subcontractor | other"
        category:
          type: string
          nullable: true
        description:
          type: string
        quantity:
          type: number
        unit:
          type: string
          nullable: true
        unitCost:
          type: number
        total:
          type: number
        materialId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - estimateId
        - type
        - description
        - quantity
        - unitCost
        - total
        - sortOrder
    CreateLineItemRequest:
      type: object
      properties:
        type:
          type: string
          default: material
        category:
          type: string
          nullable: true
        description:
          type: string
        quantity:
          type: number
        unit:
          type: string
          nullable: true
        unitCost:
          type: number
        materialId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sortOrder:
          type: integer
          default: 0
      required:
        - description
        - quantity
        - unitCost
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        estimateId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        projectName:
          type: string
        status:
          type: string
          description: "draft | sent | paid | overdue | cancelled"
        subtotal:
          type: number
        markupAmount:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        amountDue:
          type: number
        dueDate:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - projectName
        - status
        - subtotal
        - markupAmount
        - taxAmount
        - total
        - amountPaid
        - amountDue
        - createdAt
...