# lib/api-zod/src/generated/types/estimate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Estimate {
  id: number;
  estimateNumber: string;
  clientId?: number | null;
  clientName?: string | null;
  projectName: string;
  /** draft | sent | approved | rejected | invoiced */
  status: string;
  subtotal: number;
  materialsCost: number;
  laborCost: number;
  markupPercent: number;
  markupAmount: number;
  taxPercent: number;
  taxAmount: number;
  total: number;
  profitMargin: number;
  notes?: string | null;
  validUntil?: string | null;
  createdAt: string;
}