# lib/api-zod/src/generated/types/estimateDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";
import type { Estimate } from "./estimate";
import type { LineItem } from "./lineItem";

export type EstimateDetail = Estimate & {
  lineItems: LineItem[];
  client?: Client | null;
};