# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Invoice {
  id: number;
  invoiceNumber: string;
  estimateId?: number | null;
  clientId?: number | null;
  clientName?: string | null;
  projectName: string;
  /** draft | sent | paid | overdue | cancelled */
  status: string;
  subtotal: number;
  markupAmount: number;
  taxAmount: number;
  total: number;
  amountPaid: number;
  amountDue: number;
  dueDate?: string | null;
  paidAt?: string | null;
  notes?: string | null;
  createdAt: string;
}