# lib/api-zod/src/generated/types/updateEstimateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLineItemRequest } from "./createLineItemRequest";

export interface UpdateEstimateRequest {
  clientId?: number | null;
  projectName?: string;
  status?: string;
  markupPercent?: number;
  taxPercent?: number;
  notes?: string | null;
  validUntil?: string | null;
  lineItems?: CreateLineItemRequest[];
}